backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flasgger==0.9.7
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.5
openai==1.3.0
anthropic==0.8.0
google-generativeai==0.3.1
